Search in a row-wise and Colum-wise Matrix
Given a matrix of integers A of size N x M and an integer B.
In the given matrix every row and column is sorted in non-decreasing order. Find and return the position of B in the matrix in the given form:

If A[i][j] = B then return (i * 1009 + j)
If B is not present return -1.